% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knudson.R
\name{pp_knudson}
\alias{pp_knudson}
\title{Percent predicted spirometry based on Knudson (1983)}
\usage{
pp_knudson(
  sex,
  age,
  height,
  fev = NULL,
  fvc = NULL,
  vmax50 = NULL,
  vmax75 = NULL,
  fef2575 = NULL,
  ...
)
}
\arguments{
\item{sex}{Sex at birth. 1 = male, 2 = female.}

\item{age}{Age at the spirometry observation time in years.}

\item{height}{Height in centimeters.}

\item{fev}{Forced expiratory volume in one second (L)}

\item{fvc}{Forced vital capacity (L).}

\item{vmax50}{Maximal flow after exhalation of 50\% of FVC.}

\item{vmax75}{Maximal flow after exhalation of 75\% of FVC.}

\item{fef2575}{Forced expiratory flow during the middle half (25-75\%) of
FVC.}

\item{...}{Disregarded.  Input allowed so that methods which require
additional parameters (e.g. race) can be fit with consistent syntax.}
}
\description{
Calculates the percentage of the predicted spirometry variable
  (fev, fvc, etc.) using Knudson's work for the predicted values.
}
\details{
RJ Knudson, MD Lebowitz, CJ Holberg and B Burrows published
  reference spirometry equations which account for sex, age and height, but
  not race.

  The "percentage of predicted" is defined by (actual)/(predicted)*100.

  In Knudson et. al. Table 1, they state the regression coefficients.  There
  are two fits offered, with one stratifying the female fits with breaks at
  11 and 20, and the other using breaks at 11, 20 and 70.  We have used the
  simpler model without a break at 70, owing to the standards in application
  areas such as cystic fibrosis research, and Knudson's own comments that the
  model with a break at 70 might be overfitting.  For a reference on this
  being a standard in cystic fibrosis please see Table 1 in Stanojevic et.
  al. (2014), "The impact of switching..." which can be found at
  \href{http://dx.doi.org/10.1016/j.jcf.2013.11.006}{http://dx.doi.org/10.1016/j.jcf.2013.11.006}
}
