% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knudson.R
\name{predicted_knudson}
\alias{predicted_knudson}
\title{Predicted spirometry based on Knudson (1983)}
\usage{
predicted_knudson(sex, age, height, measure)
}
\arguments{
\item{sex}{Sex at birth. 1 = male, 2 = female.}

\item{age}{Age at the spirometry observation time in years.}

\item{height}{Height in centimeters.}

\item{measure}{One of 'fev', 'fvc', 'fef2575', 'vmax50', 'vmax75'.  See
\code{?pp_knudson()} for definitions of these.}
}
\description{
Calculates the predicted spirometry variable (fev, fvc, etc.)
  using Knudson's work.
}
\details{
RJ Knudson, MD Lebowitz, CJ Holberg and B Burrows published
  reference spirometry equations which account for sex, age and height, but
  not race.  DOI link: \href{https://doi.org/10.1164/arrd.1983.127.6.725}{https://doi.org/10.1164/arrd.1983.127.6.725} or PubMedID 6859656.
}
