% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spirometry.R
\docType{package}
\name{spirometry}
\alias{spirometry}
\title{\code{spirometry} package}
\description{
A unification of published prediction approaches to spirometry.
}
\details{
Currently this package only implements Knudson (1983).
}
