% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wang.R
\name{wang}
\alias{wang}
\alias{pp_wang}
\alias{predicted_wang}
\title{Wang (1993) spirometry equations.}
\usage{
pp_wang(
  sex,
  race,
  age,
  height,
  fev = NULL,
  fvc = NULL,
  fev_fvc = NULL,
  fef2575 = NULL,
  ...
)

predicted_wang(sex, age, race, height, measure)
}
\arguments{
\item{sex}{Sex at birth.  "m" = male, "f" = female.}

\item{race}{Race.  "white", "black" or "other".}

\item{age}{Age at spirometry observation (years).}

\item{height}{Height in meters.}

\item{fev}{Forced expiratory volume in one second (L)}

\item{fvc}{Forced vital capacity (L).}

\item{fev_fvc}{fev/fvc ratio.}

\item{fef2575}{Forced expiratory flow during the middle half (25-75\%) of
FVC (L/s).}

\item{...}{Disregarded.  Input allowed so that methods which require
additional parameters (e.g. race) can be fit with consistent syntax.}
}
\description{
Based on the Harvard Six Cities Study, derived for black and
  white children aged 6 to 18.
}
\details{
Regressions were derived for four measures: FVC, FEV1, FEV1/FVC and FEF25-75.
Reference: \href{https://doi.org/10.1002/ppul.1950150204}{https://doi.org/10.1002/ppul.1950150204}
}
